backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
PyMySQL==1.1.0
Pillow==11.0.0
python-decouple==3.8
django-phonenumber-field==8.1.0
pydantic==2.10.3
python-dotenv==1.0.1

# Type stubs for better IDE autocomplete
djangorestframework-stubs>=3.15.0
django-stubs>=5.2.0

#python manage.py createsuperuser --username admin --email admin@example.com